fix(tts/styletts2): pin har_source + decoder to fp32
User reported audible quality degradation after the first ~2 seconds of
synthesised audio: clean speech for the opening, then increasing
distortion/phase artefacts in the tail. Classic fp16 cumulative-precision
symptom — har_source and decoder both compute sin(2π × cumsum(f0)) in
the audio domain (har_source explicitly via SourceModuleHnNSF;
decoder via the Generator's internal SineGen). At fp16, cumsum over
74 400 samples accumulates ~10 bits of error, producing phase drift
proportional to elapsed time.

Flipping just these two stages back to fp32 restores quality (user
verified). Cost: ~287 ms extra warm decoder, ~16 ms extra har_source —
total compute ~650 ms for a 3.10 s utterance (RTFx ≈ 4.8, was 9.4 with
the broken fp16 tail). Still well below real-time and well above the
all-fp32 baseline (~2.9 s compute).

Final per-stage manifest:

  text_encoder       fp16  CPU_ONLY
  bert               fp32  CPU_AND_NE
  ref_encoder        fp16  CPU_AND_NE
  diffusion_unet     fp16  CPU_AND_NE
  duration_predictor fp16  CPU_ONLY
  f0n_predictor      fp16  CPU_AND_NE
  har_source         fp32  CPU_AND_GPU   <- was fp16
  decoder            fp32  CPU_ONLY      <- was fp16

+++ models/tts/styletts2/coreml/inference.py
@@ -111,8 +111,13 @@ _STAGE_PRECISION: dict[str, str] = {
     "diffusion_unet":     "fp16",
     "duration_predictor": "fp16",
     "f0n_predictor":      "fp16",
-    "har_source":         "fp16",
-    "decoder":            "fp16",
+    # har_source / decoder kept in fp32: both compute sin(2π × cumsum(f0))
+    # in the audio domain (har_source explicitly, decoder inside SineGen).
+    # fp16 cumsum loses ~10 bits of precision after a few thousand samples,
+    # producing audible phase drift that gets worse the further into the
+    # clip you go (clean first ~2 s, distorted tail).
+    "har_source":         "fp32",
+    "decoder":            "fp32",
 }
